Fixed and tested report creation and retrieval. Moved creation to Monday morning.

+++ goci-core/goci-repository/src/main/java/uk/ac/ebi/spot/goci/repository/WeeklyReportRepository.java
@@ -9,5 +9,5 @@ import java.util.List;
 @RepositoryRestResource(exported = false)
 public interface WeeklyReportRepository extends JpaRepository<WeeklyReport, Long> {
 
-    List<WeeklyReport> findByTypeAndWeekCode(String type, int weekCode);
+    List<WeeklyReport> findByTypeAndWeekCode(String type, Long weekCode);
 }
